pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mavetools"
dynamic = ["version"]
description = "Useful functions for manipulating Multiplex Assay of Variant Effect datasets."
readme = "README.md"
license = ""
requires-python = ">=3.7"
authors = [
    { name = "MaveDB Developers", email = "alan.rubin@wehi.edu.au" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: BSD License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "aiohttp",
    "attrs",
    "certifi",
    "fqfa>=1.2.1",
    "idutils",
    "mavehgvs",
    "numpy",
    "pandas",
    "requests",
    "tables",
    "joblib",
    "mavedb>=2023.1",
    "pyhumps",
]

[project.urls]
repository = "https://github.com/VariantEffect/mavetools"
documentation = "https://mavedb.org/docs/mavetools/"

[project.optional-dependencies]
dev = [
    "black",
    "flake8",
    "pre-commit",
    "pytest",
]

[tool.black]
line-length = 120

[tool.hatch.version]
path = "src/mavetools/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/mavetools"]

[tool.hatch.build.targets.sdist]
exclude = [
    "docs",
    ".github",
]
